+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# [3.32.0](https://github.com/samtun/hackworld/compare/v3.31.0...v3.32.0) (2026-06-28)
+
+
+### Features
+
+* fix fade in and out opacity settings ([df285e9](https://github.com/samtun/hackworld/commit/df285e958fd86af20e4d1b559489139e98223ad6))
+* show controller mapping in a popup instead of new tab ([76a3f49](https://github.com/samtun/hackworld/commit/76a3f49a4da2e29cb32c31ae7ea3518551a83e33))
+
+# [3.31.0](https://github.com/samtun/hackworld/compare/v3.30.4...v3.31.0) (2026-06-28)
+
+
+### Features
+
+* add tests for particles of healing station ([543afb8](https://github.com/samtun/hackworld/commit/543afb80fba0f54212e783d813714fa06a822425))
+* add tests for teleporter ([9ad6c3a](https://github.com/samtun/hackworld/commit/9ad6c3a5937acca098c5b09abd7734e576a19e99))
+* decrease particle size when not healing ([27c4fb2](https://github.com/samtun/hackworld/commit/27c4fb26f93db65d1227c32640c8fb05f91c276a))
+* increase particle size and speed when player is nearby ([ffe16ee](https://github.com/samtun/hackworld/commit/ffe16eea6108313984d96986be9ffc509edaa79e))
+
 ## [3.30.4](https://github.com/samtun/hackworld/compare/v3.30.3...v3.30.4) (2026-06-28)